fix(uikit): support compatibility for removing AppState listener under 0.65
