+++ packages/uikit-utils/package.json
@@ -43,6 +43,7 @@
   },
   "devDependencies": {
     "@types/react": "*",
+    "@types/react-native": "*",
     "react": "17.0.2",
     "react-native": "0.67.4",
     "react-native-builder-bob": "^0.18.0",
