+++ packages/uikit-utils/src/index.ts
@@ -22,6 +22,7 @@ export { BufferedRequest } from './shared/bufferedRequest';
 export * from './shared';
 
 export * from './hooks';
+export * from './hooks/react-native';
 export * from './ui-format/groupChannel';
 export * from './ui-format/common';
 export * from './sendbird/channel';
